requirements.txt
setuptools
mysqlclient
mysql-replication
clickhouse-driver
configobj
clickhouse-driver